+++ youwol/configuration/config_from_static_file.py
@@ -4,7 +4,7 @@ from typing import Optional, Dict
 from youwol.configuration.configuration_handler import ConfigurationHandler
 from youwol.configuration.models_config import Profiles, ConfigurationData, Cascade, CascadeBaseProfile, \
     ExtendingProfile
-from youwol.utils_paths import app_dirs, PathException, existing_path_or_default
+from youwol_utils.utils_paths import app_dirs, PathException, existing_path_or_default
 
 
 class ConfigurationDataWithCascade(ConfigurationData):
