+++ youwol/routers/environment/upload_assets/story.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from youwol.environment.youwol_environment import YouwolEnvironment
 from youwol.environment.clients import RemoteClients
 from youwol.routers.environment.upload_assets.models import UploadTask
-from youwol.utils_paths import parse_json, write_json
+from youwol_utils.utils_paths import parse_json, write_json
 from youwol_utils import JSON
 from youwol_utils.clients.assets_gateway.assets_gateway import AssetsGatewayClient
